+++ driver-core/src/test/java/com/datastax/driver/core/ReconnectionTest.java
@@ -1,18 +1,22 @@
 package com.datastax.driver.core;
 
+import java.io.IOException;
 import java.net.InetSocketAddress;
+import java.util.Collections;
+import java.util.Iterator;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 
+import static java.util.concurrent.TimeUnit.SECONDS;
+
 import com.google.common.base.Charsets;
 import com.google.common.collect.ImmutableMap;
 import org.testng.annotations.Test;
-
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.datastax.driver.core.policies.ConstantReconnectionPolicy;
-import com.datastax.driver.core.policies.ReconnectionPolicy;
+import com.datastax.driver.core.Host.State;
+import com.datastax.driver.core.policies.*;
 
 import static com.datastax.driver.core.Assertions.assertThat;
 
@@ -21,15 +25,16 @@ import static com.datastax.driver.core.Assertions.assertThat;
  */
 public class ReconnectionTest {
 
-    @Test(groups = "short")
-    public void should_reconnect_after_full_connectivity_lost() {
+    @Test(groups = "long")
+    public void should_reconnect_after_full_connectivity_loss() throws InterruptedException {
         CCMBridge ccm = null;
         Cluster cluster = null;
         try {
             ccm = CCMBridge.create("test", 2);
+            int reconnectionDelay = 1000;
             cluster = Cluster.builder()
                              .addContactPoint(CCMBridge.ipOfNode(1))
-                             .withReconnectionPolicy(new ConstantReconnectionPolicy(1000))
+                             .withReconnectionPolicy(new ConstantReconnectionPolicy(reconnectionDelay))
                              .build();
             cluster.connect();
 
@@ -44,7 +49,10 @@ public class ReconnectionTest {
             ccm.start(2);
             ccm.waitForUp(2);
 
-            TestUtils.waitFor(CCMBridge.ipOfNode(2), cluster);
+            assertThat(cluster).host(2).comesUpWithin(120, SECONDS);
+
+            // Give the control connection a few moments to reconnect
+            TimeUnit.MILLISECONDS.sleep(reconnectionDelay * 2);
             assertThat(cluster).usesControlHost(2);
         } finally {
             if (cluster != null)
@@ -54,7 +62,7 @@ public class ReconnectionTest {
         }
     }
 
-    @Test(groups = "short")
+    @Test(groups = "long")
     public void should_keep_reconnecting_on_authentication_error() throws InterruptedException {
         CCMBridge ccm = null;
         Cluster cluster = null;
@@ -70,11 +78,11 @@ public class ReconnectionTest {
             CountingReconnectionPolicy reconnectionPolicy = new CountingReconnectionPolicy(new ConstantReconnectionPolicy(reconnectionDelayMs));
 
             cluster = Cluster.builder()
-                            .addContactPoint(CCMBridge.ipOfNode(1))
-                            // Start with the correct auth so that we can initialize the server
-                            .withAuthProvider(authProvider)
-                            .withReconnectionPolicy(reconnectionPolicy)
-                            .build();
+                             .addContactPoint(CCMBridge.ipOfNode(1))
+                             // Start with the correct auth so that we can initialize the server
+                             .withAuthProvider(authProvider)
+                             .withReconnectionPolicy(reconnectionPolicy)
+                             .build();
 
             cluster.init();
             assertThat(cluster).usesControlHost(1);
@@ -100,7 +108,104 @@ public class ReconnectionTest {
             authProvider.password = "cassandra";
 
             // The driver should eventually reconnect to the node
-            TestUtils.waitFor(CCMBridge.ipOfNode(1), cluster);
+            assertThat(cluster).host(1).comesUpWithin(120, SECONDS);
+        } finally {
+            if (cluster != null)
+                cluster.close();
+            if (ccm != null)
+                ccm.remove();
+        }
+    }
+
+    @Test(groups = "long")
+    public void should_cancel_reconnection_attempts() throws InterruptedException {
+        CCMBridge ccm = null;
+        Cluster cluster = null;
+
+        long reconnectionDelayMillis = 1 * 1000;
+        CountingReconnectionPolicy reconnectionPolicy = new CountingReconnectionPolicy(new ConstantReconnectionPolicy(reconnectionDelayMillis));
+
+        try {
+            ccm = CCMBridge.create("test", 2);
+            cluster = Cluster.builder()
+                             .addContactPoint(CCMBridge.ipOfNode(1))
+                             .withReconnectionPolicy(reconnectionPolicy)
+                             .build();
+            cluster.connect();
+
+            // Stop a node and cancel the reconnection attempts to it
+            ccm.stop(2);
+            Host host2 = TestUtils.findHost(cluster, 2);
+            host2.getReconnectionAttemptFuture().cancel(false);
+
+            // The reconnection count should not vary over time anymore
+            int initialCount = reconnectionPolicy.count.get();
+            TimeUnit.MILLISECONDS.sleep(reconnectionDelayMillis * 2);
+            assertThat(reconnectionPolicy.count.get()).isEqualTo(initialCount);
+
+            // Restart the node, which will trigger an UP notification
+            ccm.start(2);
+            ccm.waitForUp(2);
+
+            // The driver should now see the node as UP again
+            assertThat(cluster).host(2).comesUpWithin(120, SECONDS);
+
+        } finally {
+            if (cluster != null)
+                cluster.close();
+            if (ccm != null)
+                ccm.remove();
+        }
+    }
+
+    @Test(groups = "long")
+    public void should_trigger_one_time_reconnect() throws InterruptedException, IOException {
+        CCMBridge ccm = null;
+        Cluster cluster = null;
+
+        long reconnectionDelayMillis = 1 * 1000;
+        TogglabePolicy loadBalancingPolicy = new TogglabePolicy(new RoundRobinPolicy());
+
+        try {
+            ccm = CCMBridge.create("test", 1);
+            cluster = Cluster.builder()
+                             .addContactPoint(CCMBridge.ipOfNode(1))
+                             .withLoadBalancingPolicy(loadBalancingPolicy)
+                             .withReconnectionPolicy(new ConstantReconnectionPolicy(reconnectionDelayMillis))
+                             .build();
+            cluster.connect();
+
+            // Tweak the LBP so that the control connection never reconnects, otherwise
+            // it would interfere with the rest of the test (this is a bit of a hack)
+            loadBalancingPolicy.returnEmptyQueryPlan = true;
+
+            // Stop the node and cancel the reconnection attempts to it
+            ccm.stop(1);
+            ccm.waitForDown(1);
+            assertThat(cluster).host(1).goesDownWithin(20, SECONDS);
+            Host host1 = TestUtils.findHost(cluster, 1);
+            host1.getReconnectionAttemptFuture().cancel(false);
+
+            // Trigger a one-time reconnection attempt (this will fail)
+            host1.tryReconnectOnce();
+
+            // Wait for a few reconnection cycles before checking
+            TimeUnit.MILLISECONDS.sleep(reconnectionDelayMillis * 2);
+            assertThat(cluster).host(1).hasState(State.DOWN);
+
+            // Restart the node (this will not trigger an UP notification thanks to our
+            // hack to disable the control connection reconnects). The host should stay
+            // down for the driver.
+            ccm.start(1);
+            ccm.waitForUp(1);
+
+            TimeUnit.SECONDS.sleep(Cluster.NEW_NODE_DELAY_SECONDS);
+            assertThat(cluster).host(1).hasState(State.DOWN);
+
+            // Trigger another one-time reconnection attempt (this will succeed). The
+            // host should be back up.
+            host1.tryReconnectOnce();
+            assertThat(cluster).host(1).comesUpWithin(120, SECONDS);
         } finally {
             if (cluster != null)
                 cluster.close();
@@ -189,6 +294,26 @@ public class ReconnectionTest {
                 count.incrementAndGet();
                 return childSchedule.nextDelayMs();
             }
+        }
+    }
+
+    /**
+     * A load balancing policy that can be "disabled" by having its query plan return no hosts.
+     */
+    public static class TogglabePolicy extends DelegatingLoadBalancingPolicy {
+
+        volatile boolean returnEmptyQueryPlan;
+
+        public TogglabePolicy(LoadBalancingPolicy delegate) {
+            super(delegate);
+        }
+
+        @Override
+        public Iterator<Host> newQueryPlan(String loggedKeyspace, Statement statement) {
+            if (returnEmptyQueryPlan)
+                return Collections.<Host>emptyList().iterator();
+            else
+                return super.newQueryPlan(loggedKeyspace, statement);
         }
     }
 }